The user profile has been refactored and implements the skill rank, with pseudo ACL check.

+++ src/main/java/auh/helper/Skill.java
@@ -42,8 +42,8 @@ public class Skill {
                 };
         }
 
-        public static double getWeightedScore(Integer skillValue){
-                return (1-(1/(Math.pow(skillValue.doubleValue(),0.5))));
+        public static double getWeightedScore(Double skillValue){
+                return (1-(1/(Math.pow(skillValue,0.5))));
         }
